// src/models/Category.ts
/**
 * Categorias disponíveis para itens de closet
 */
export enum Category {
  TOPS = 'tops',
  BOTTOMS = 'bottoms',
  DRESSES = 'dresses',
  OUTERWEAR = 'outerwear',
  SHOES = 'shoes',
  ACCESSORIES = 'accessories',
  BAGS = 'bags',
  JEWELRY = 'jewelry',
  UNDERWEAR = 'underwear',
  ACTIVEWEAR = 'activewear',
  SLEEPWEAR = 'sleepwear',
  OTHER = 'other'
}

/**
 * Cores disponíveis para itens
 */
export enum Color {
  BLACK = 'black',
  WHITE = 'white',
  GRAY = 'gray',
  RED = 'red',
  BLUE = 'blue',
  GREEN = 'green',
  YELLOW = 'yellow',
  ORANGE = 'orange',
  PURPLE = 'purple',
  PINK = 'pink',
  BROWN = 'brown',
  BEIGE = 'beige',
  MULTICOLOR = 'multicolor',
  OTHER = 'other'
}

/**
 * Estações do ano
 */
export enum Season {
  SPRING = 'spring',
  SUMMER = 'summer',
  FALL = 'fall',
  WINTER = 'winter',
  ALL_SEASONS = 'all_seasons'
}